Add a dfs flag for scanning xrootd sites with distributed filesystem
To list files on a local redirector, XRootD will query every registered data server. But for sites running a distributed file system (`cms.dfs`), the data servers have an identical view of the data, and it's only necessary to query one.

For sites flagged `dfs: true`, query the local redirector and choose a single data server to scan.